--- EventView-1/src/main/java/com/eventview/controller/LoggingController.java
@@ -1,23 +1,0 @@
-package com.eventview.controller;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class LoggingController {
-
-    private Logger logger = LoggerFactory.getLogger(LoggingController.class);
-
-    @RequestMapping("/")
-    public String index() {
-        logger.trace("A TRACE Message");
-        logger.debug("A DEBUG Message");
-        logger.info("An INFO Message");
-        logger.warn("A WARN Message");
-        logger.error("An ERROR Message");
-
-        return "See the logs...";
-    }
-}
